+++ server/src/routes/user.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { verifyUser } from 'src/middleware/auth.middleware';
+import { verifyUser } from 'src/utils/jwt.util';
 import { users } from 'src/controllers/user.controller';
 
 const userRouter = Router();
