+++ server/src/services/auth.service.ts
@@ -1,8 +1,4 @@
-import {
-    getRefreshToken,
-    getToken,
-    verifyToken,
-} from 'src/middleware/auth.middleware';
+import { getRefreshToken, getToken, verifyToken } from 'src/utils/jwt.util';
 import User, { IUser } from 'src/model/user.model';
 
 export interface UserSingUpInfo {
